chore: review table
