+++ packages/code-editor/src/CodeEditor/CodeEditor.stories.tsx
@@ -5,13 +5,9 @@ import {
 } from "@hitachivantara/uikit-react-code-editor";
 
 import { CustomPluginStory } from "./stories/CustomPlugin";
-import CustomPluginStoryRaw from "./stories/CustomPlugin/CustomPlugin?raw";
 import { MainStory } from "./stories/Main";
-import MainStoryRaw from "./stories/Main?raw";
 import { SqlStory } from "./stories/Sql";
-import SqlStoryRaw from "./stories/Sql/Sql?raw";
 import { XmlStory } from "./stories/Xml";
-import XmlStoryRaw from "./stories/Xml/Xml?raw";
 import { setupChromatic } from ".storybook/setupChromatic";
 
 const meta: Meta<typeof HvCodeEditor> = {
@@ -22,9 +18,6 @@ export default meta;
 
 export const Main: StoryObj<HvCodeEditorProps> = {
   parameters: {
-    docs: {
-      source: { code: MainStoryRaw },
-    },
     ...setupChromatic(["DS5 dawn"], 5000),
   },
   render: () => <MainStory />,
@@ -55,7 +48,6 @@ export const XMLEditor: StoryObj<HvCodeEditorProps> = {
         story:
           "XML is one of the languages supported, and it can be enabled by setting the `language` property to `xml`. A XML schema can also be provided through the `schema` property. By providing a XML schema, the XML written will be validated against the schema showing errors. Providing a schema will also enable the code editor to show suggestions when opening a tag (`<`), writing an attribute, and when clicking on the CTRL and SPACE keys at the same time. By default, the XML code editor is formatted automatically. The property `disableAutoFormat` can be set to `true` to disable this behavior. You can also format manually the code using the `hvXmlFormatter` util.",
       },
-      source: { code: XmlStoryRaw },
     },
   },
   render: () => <XmlStory />,
@@ -68,7 +60,6 @@ export const SQLEditor: StoryObj<HvCodeEditorProps> = {
         story:
           "SQL is one of the languages supported, and it can be enabled by setting the `language` property to `sql`. A SQL schema can also be provided through the `schema` property. Providing a SQL schema will enable the code editor to show suggestions when writing the query, and when clicking on the CTRL and SPACE keys at the same time.",
       },
-      source: { code: SqlStoryRaw },
     },
   },
   render: () => <SqlStory />,
@@ -81,7 +72,6 @@ export const CustomPlugin: StoryObj<HvCodeEditorProps> = {
         story:
           "A Code Editor with a custom language plugin that extends the Code Editor's own XML language plugin to show different suggestions and format the code differently.",
       },
-      source: { code: CustomPluginStoryRaw },
     },
   },
   render: () => <CustomPluginStory />,
