--- packages/core/src/Table/stories/TableRenderers/DropdownColumnRenderer.tsx
@@ -1,116 +1,0 @@
-import { useMemo, useState } from "react";
-import {
-  hvDropdownColumn,
-  HvPagination,
-  HvRowInstance,
-  HvTable,
-  HvTableBody,
-  HvTableCell,
-  HvTableContainer,
-  HvTableHead,
-  HvTableHeader,
-  HvTableRow,
-  useHvData,
-  useHvPagination,
-} from "@hitachivantara/uikit-react-core";
-
-import { EmptyRow, makeRenderersData, NewRendererEntry } from "./utils";
-
-export const DropdownColumnRenderer = () => {
-  const [data, setData] = useState(() => makeRenderersData(64));
-
-  const columns = useMemo(() => {
-    return [
-      hvDropdownColumn<NewRendererEntry, string>(
-        { Header: "Severity", accessor: "severity", id: "severity-header" },
-        undefined,
-        "Select severity...",
-        "Select severity...",
-        (id, value) => {
-          let newData = [...data];
-          newData = newData.map((val, index) => {
-            const newVal = { ...val };
-            if (index.toString() === id) {
-              if (newVal.severity) {
-                newVal.severity = newVal.severity.map((sev) => {
-                  const newSev = { ...sev };
-                  newSev.selected = false;
-                  if (newSev.id === value?.id) newSev.selected = value.selected;
-                  return newSev;
-                });
-              }
-            }
-            return newVal;
-          });
-          setData(newData);
-        },
-      ),
-    ];
-  }, [data]);
-
-  const {
-    getTableProps,
-    getTableBodyProps,
-    prepareRow,
-    headers,
-    page,
-    getHvPaginationProps,
-  } = useHvData<NewRendererEntry, string>(
-    {
-      initialState: { pageSize: 5 },
-      columns,
-      data,
-    },
-    useHvPagination,
-  );
-
-  const rowRenderer = (pages: HvRowInstance<NewRendererEntry, string>[]) => {
-    return pages.map((row, i) => {
-      prepareRow(row);
-      const { key, ...rowProps } = row.getRowProps({ "aria-rowindex": i + 1 });
-
-      return (
-        <HvTableRow key={key} {...rowProps}>
-          {row.cells.map((cell) => (
-            <HvTableCell {...cell.getCellProps()} key={cell.getCellProps().key}>
-              {cell.render("Cell")}
-            </HvTableCell>
-          ))}
-        </HvTableRow>
-      );
-    });
-  };
-
-  return (
-    <>
-      <HvTableContainer>
-        <HvTable
-          {...getTableProps({
-            "aria-rowcount": data.length,
-          })}
-          style={{
-            width: 230,
-          }}
-        >
-          <HvTableHead>
-            <HvTableRow>
-              {headers.map((col) => (
-                <HvTableHeader
-                  {...col.getHeaderProps()}
-                  key={col.Header}
-                  id={col.id}
-                >
-                  {col.render("Header")}
-                </HvTableHeader>
-              ))}
-            </HvTableRow>
-          </HvTableHead>
-          <HvTableBody {...getTableBodyProps()}>
-            {page.length === 0 ? <EmptyRow height={100} /> : rowRenderer(page)}
-          </HvTableBody>
-        </HvTable>
-      </HvTableContainer>
-      <HvPagination {...getHvPaginationProps?.()} />
-    </>
-  );
-};
