--- packages/core/src/Table/stories/TableRenderers/ProgressColumnRenderer.tsx
@@ -1,103 +1,0 @@
-import { useMemo, useState } from "react";
-import {
-  HvPagination,
-  hvProgressColumn,
-  HvRowInstance,
-  HvTable,
-  HvTableBody,
-  HvTableCell,
-  HvTableContainer,
-  HvTableHead,
-  HvTableHeader,
-  HvTableRow,
-  useHvData,
-  useHvPagination,
-} from "@hitachivantara/uikit-react-core";
-
-import { EmptyRow, makeRenderersData, NewRendererEntry } from "./utils";
-
-export const ProgressColumnRenderer = () => {
-  const columns = useMemo(() => {
-    return [
-      hvProgressColumn<NewRendererEntry, string>(
-        {
-          Header: "Probability",
-          accessor: "riskScore",
-          style: { width: 125 },
-          disableSortBy: true,
-          id: "probability-header",
-        },
-        (row) => row.original.riskScore,
-        () => 100,
-        "secondary",
-      ),
-    ];
-  }, []);
-
-  const [data] = useState(() => makeRenderersData(64));
-
-  const {
-    getTableProps,
-    getTableBodyProps,
-    prepareRow,
-    headers,
-    page,
-    getHvPaginationProps,
-  } = useHvData<NewRendererEntry, string>(
-    {
-      columns,
-      data,
-    },
-    useHvPagination,
-  );
-
-  const rowRenderer = (pages: HvRowInstance<NewRendererEntry, string>[]) => {
-    return pages.map((row, i) => {
-      prepareRow(row);
-      const { key, ...rowProps } = row.getRowProps({ "aria-rowindex": i + 1 });
-
-      return (
-        <HvTableRow key={key} {...rowProps}>
-          {row.cells.map((cell) => (
-            <HvTableCell {...cell.getCellProps()} key={cell.getCellProps().key}>
-              {cell.render("Cell")}
-            </HvTableCell>
-          ))}
-        </HvTableRow>
-      );
-    });
-  };
-
-  return (
-    <>
-      <HvTableContainer>
-        <HvTable
-          {...getTableProps({
-            "aria-rowcount": data.length,
-          })}
-          style={{
-            width: 230,
-          }}
-        >
-          <HvTableHead>
-            <HvTableRow>
-              {headers.map((col) => (
-                <HvTableHeader
-                  {...col.getHeaderProps()}
-                  key={col.Header}
-                  id={col.id}
-                >
-                  {col.render("Header")}
-                </HvTableHeader>
-              ))}
-            </HvTableRow>
-          </HvTableHead>
-          <HvTableBody {...getTableBodyProps()}>
-            {page.length === 0 ? <EmptyRow height={100} /> : rowRenderer(page)}
-          </HvTableBody>
-        </HvTable>
-      </HvTableContainer>
-      <HvPagination {...getHvPaginationProps?.()} />
-    </>
-  );
-};
