--- packages/core/src/Table/stories/TableRenderers/TextColumnRenderer.tsx
@@ -1,93 +1,0 @@
-import { useMemo, useState } from "react";
-import {
-  HvPagination,
-  HvRowInstance,
-  HvTable,
-  HvTableBody,
-  HvTableCell,
-  HvTableContainer,
-  HvTableHead,
-  HvTableHeader,
-  HvTableRow,
-  hvTextColumn,
-  useHvData,
-  useHvPagination,
-} from "@hitachivantara/uikit-react-core";
-
-import { EmptyRow, makeRenderersData, NewRendererEntry } from "./utils";
-
-export const TextColumnRenderer = () => {
-  const columns = useMemo(
-    () => [
-      hvTextColumn<NewRendererEntry, string>({
-        Header: "Event Type",
-        accessor: "eventType",
-        style: { maxWidth: 160 },
-      }),
-    ],
-    [],
-  );
-
-  const [data] = useState(() => makeRenderersData(64));
-
-  const {
-    getTableProps,
-    getTableBodyProps,
-    prepareRow,
-    headers,
-    page,
-    getHvPaginationProps,
-  } = useHvData<NewRendererEntry, string>(
-    {
-      columns,
-      data,
-    },
-    useHvPagination,
-  );
-
-  const rowRenderer = (pages: HvRowInstance<NewRendererEntry, string>[]) => {
-    return pages.map((row, i) => {
-      prepareRow(row);
-      const { key, ...rowProps } = row.getRowProps({ "aria-rowindex": i + 1 });
-
-      return (
-        <HvTableRow key={key} {...rowProps}>
-          {row.cells.map((cell) => (
-            <HvTableCell {...cell.getCellProps()} key={cell.getCellProps().key}>
-              {cell.render("Cell")}
-            </HvTableCell>
-          ))}
-        </HvTableRow>
-      );
-    });
-  };
-
-  return (
-    <>
-      <HvTableContainer>
-        <HvTable
-          {...getTableProps({
-            "aria-rowcount": data.length,
-          })}
-          style={{
-            width: 230,
-          }}
-        >
-          <HvTableHead>
-            <HvTableRow>
-              {headers.map((col) => (
-                <HvTableHeader {...col.getHeaderProps()} key={col.Header}>
-                  {col.render("Header")}
-                </HvTableHeader>
-              ))}
-            </HvTableRow>
-          </HvTableHead>
-          <HvTableBody {...getTableBodyProps()}>
-            {page.length === 0 ? <EmptyRow height={100} /> : rowRenderer(page)}
-          </HvTableBody>
-        </HvTable>
-      </HvTableContainer>
-      <HvPagination {...getHvPaginationProps?.()} />
-    </>
-  );
-};
